Serialise: Use long double for float serializations

+++ src/serialise.h
@@ -124,7 +124,7 @@ namespace Serialise {
 	 * Serialise field_value according to field_type.
 	 */
 
-	std::string _float(FieldType field_type, double field_value);
+	std::string _float(FieldType field_type, long double field_value);
 	std::string integer(FieldType field_type, std::int64_t field_value);
 	std::string positive(FieldType field_type, std::uint64_t field_value);
 	std::string boolean(FieldType field_type, bool field_value);
@@ -169,7 +169,7 @@ namespace Serialise {
 	// Serialise field_value like float.
 	std::string _float(std::string_view field_value);
 
-	inline std::string _float(double field_value) {
+	inline std::string _float(long double field_value) {
 		return sortable_serialise(field_value);
 	}
 
@@ -250,7 +250,7 @@ namespace Serialise {
 		return boolean(val);
 	}
 
-	inline std::string serialise(double val) {
+	inline std::string serialise(long double val) {
 		return _float(val);
 	}
 
@@ -289,7 +289,7 @@ namespace Unserialise {
 	double timedelta_d(std::string_view serialised_time);
 
 	// Unserialise a serialised float.
-	inline double _float(std::string_view serialised_float) {
+	inline long double _float(std::string_view serialised_float) {
 		return sortable_unserialise(serialised_float);
 	}
